feat: Moved download logic into own service
